requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
tensorflow>=2.8.0
dwave-ocean-sdk>=4.0.0
blpapi>=3.24.10
